# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PriceCheck NG API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors with optional location filter
      parameters:
        - name: location
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Register a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID with ratings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor with rating info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdateInput"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /vendors/{id}/products:
    get:
      operationId: getVendorProducts
      tags: [vendors]
      summary: Get all products and prices for a vendor (vendor dashboard)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor product listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorProductListing"
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, location]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        address:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        verified:
          type: boolean
        whatsapp:
          type: ["string", "null"]
        stock_available:
          type: boolean
        delivery_options:
          type: ["string", "null"]
        response_time:
          type: ["string", "null"]
        price_count:
          type: ["integer", "null"]
        average_rating:
          type: ["number", "null"]
        rating_count:
          type: ["integer", "null"]
    VendorDetail:
      type: object
      required: [id, name, location]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        address:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        verified:
          type: boolean
        whatsapp:
          type: ["string", "null"]
        stock_available:
          type: boolean
        delivery_options:
          type: ["string", "null"]
        response_time:
          type: ["string", "null"]
        price_count:
          type: ["integer", "null"]
        average_rating:
          type: ["number", "null"]
        rating_count:
          type: ["integer", "null"]
        recent_reviews:
          type: array
          items:
            $ref: "#/components/schemas/VendorReview"
    VendorInput:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string
        address:
          type: string
        logo_url:
          type: string
        whatsapp:
          type: string
        delivery_options:
          type: string
        response_time:
          type: string
    VendorUpdateInput:
      type: object
      properties:
        stock_available:
          type: boolean
        delivery_options:
          type: string
        response_time:
          type: string
        whatsapp:
          type: string
        address:
          type: string
    VendorProductListing:
      type: object
      required: [price_id, product_id, product_name, unit, price, quantity, updated_at]
      properties:
        price_id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        category_name:
          type: ["string", "null"]
        unit:
          type: string
        price:
          type: number
        quantity:
          type: string
        updated_at:
          type: string
        stock_available:
          type: boolean
    VendorReview:
      type: object
      required: [id, vendor_id, rating, reviewer_name, created_at]
      properties:
        id:
          type: integer
        vendor_id:
          type: integer
        rating:
          type: integer
        reviewer_name:
          type: string
        comment:
          type: ["string", "null"]
        created_at:
          type: string
...