# lib/api-zod/src/generated/types/vendorDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VendorReview } from './vendorReview';

export interface VendorDetail {
  id: number;
  name: string;
  location: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  logo_url?: string | null;
  verified?: boolean;
  /** @nullable */
  whatsapp?: string | null;
  stock_available?: boolean;
  /** @nullable */
  delivery_options?: string | null;
  /** @nullable */
  response_time?: string | null;
  /** @nullable */
  price_count?: number | null;
  /** @nullable */
  average_rating?: number | null;
  /** @nullable */
  rating_count?: number | null;
  recent_reviews?: VendorReview[];
}