# lib/api-zod/src/generated/types/vendorProductListing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VendorProductListing {
  price_id: number;
  product_id: number;
  product_name: string;
  /** @nullable */
  category_name?: string | null;
  unit: string;
  price: number;
  quantity: string;
  updated_at: string;
  stock_available?: boolean;
}